requirements.txt
eventlet==0.40.0
flask==3.1.1
flask-socketio==5.5.1
protobuf==6.30.1
pyzmq==26.2.1